refactor: provider를 관리하는 lib 폴더의 위치를 app 폴더 내부로 이동
