--- src/lib/index.tsx
@@ -1,1 +1,0 @@
-export { default as RecoilProvider } from './RecoilProvider';
